deleteAll is now recursive
As we started using this method, we realized that we always want all
child scopes to also be cleared. Child scopes are primarily used for
logged in users or for grouping a full feature together.

Added a test case to ensure the existing behavior remains in addition to
the recursion.