Bump versions to address build errors
Prior to this commit, opening this project on IntelliJ IDEA 2022.2.3
will result in a complaint that the gradle wrapper is too old.

After bumping that to 7.3.3, ``./gradlew build` fails with mavenPublish
errors - presumably due to the version is not compatible with the
new gradle version.

After bumping mavenPublish to 0.22.0 - the build fails with incompatible
android plugin versions.

After updating android gradle plugin to 7.2.2 - it then complains about
some annotation usage in errorProne and nullAway plugins

After updating those, we are hit with a dependency error because the
protobuf version we use does not have binaries for osx arm machines
(which is the more popular osx developer machine for potential
contributors of this library) - so we also bumped it to 3.22.3

As the spiritual successor to https://github.com/uber/simple-store/pull/88 - this commit also bumped the kotlin
version to match