Separate the "twice" test case
In 0eb6c51, we changed `flush` to be an AtomicReference, but that alone
is not enough to have consistent results since the old version of the
"twice" test still may or may not interweave the setting of `flush`
between the two consecutive calls.

This commit splits the two tests to assert for both the sequential and
the non-sequential case.